requirements.txt
flask
mysql-connector-python
flask-mysqldb
python-dotenv